Sort Jupyter-developed extensions to the top.
