+++ packages/extensionmanager/src/model.ts
@@ -17,7 +17,7 @@ import {
 
 import { reportInstallError } from './dialog';
 
-import { Searcher, ISearchResult } from './query';
+import { Searcher, ISearchResult, isJupyterOrg } from './query';
 
 /**
  * Information about an extension.
@@ -344,7 +344,7 @@ export class ListModel extends VDomModel {
     for (let key of Object.keys(installedMap)) {
       installed.push(installedMap[key]);
     }
-    this._installed = installed;
+    this._installed = installed.sort(Private.comparator);
 
     let searchResult: IEntry[] = [];
     for (let key of Object.keys(searchMap)) {
@@ -354,7 +354,7 @@ export class ListModel extends VDomModel {
         searchResult.push(installedMap[key]);
       }
     }
-    this._searchResult = searchResult;
+    this._searchResult = searchResult.sort(Private.comparator);
     try {
       this._totalEntries = (await search).total;
     } catch (error) {
@@ -678,4 +678,29 @@ function handleError(response: Response): Response {
     throw new Error(`${response.status} (${response.statusText})`);
   }
   return response;
+}
+
+/**
+ * A namespace for private functionality.
+ */
+namespace Private {
+  /**
+   * A comparator function that sorts whitelisted orgs to the top.
+   */
+  export function comparator(a: IEntry, b: IEntry): number {
+    if (a.name === b.name) {
+      return 0;
+    }
+
+    let testA = isJupyterOrg(a.name);
+    let testB = isJupyterOrg(b.name);
+
+    if (testA === testB) {
+      return a.name.localeCompare(b.name);
+    } else if (testA && !testB) {
+      return -1;
+    } else {
+      return 1;
+    }
+  }
 }
