+++ packages/extensionmanager/src/query.ts
@@ -282,3 +282,21 @@ export class Searcher {
    */
   cdnUri: string;
 }
+
+/**
+ * Check whether the NPM org is a Jupyter one.
+ */
+export function isJupyterOrg(name: string): boolean {
+  /**
+   * A list of whitelisted NPM orgs.
+   */
+  const whitelist = ['jupyterlab', 'jupyter-widgets'];
+  const parts = name.split('/');
+  const first = parts[0];
+  return (
+    parts.length > 1 && // Has a first part
+    first && // with a finite length
+    first[0] === '@' && // corresponding to an org name
+    whitelist.indexOf(first.slice(1)) !== -1 // in the org whitelist.
+  );
+}
